Add files via upload
The Online Exam System project aims to create a user-friendly application for conducting online 
exams. It utilizes Java Swing for the graphical user interface (GUI) and does not rely on a database for storing 
data. The project provides a platform for students to take exams online and for teachers to create and manage 
exams. 
Key Features: 
In this project, there are given 10 questions to play. User can bookmark any question for the 
reconsideration while result going to resul